Support_Chatbot/requirements.txt
langchain_openai 
langchain_core
python-dotenv
streamlit
chainlit
requests